requirements.txt
aiofiles==24.1.0
fastapi==0.95.0
httpx==0.27.2
pymongo==4.6.0
python-dotenv==1.0.1
tenacity==9.0.0

